refactor: add twitter callback route
